Split RowRef.cs, make SubrowRef return a collection
